+++ emanote/src/Emanote.hs
@@ -22,6 +22,7 @@ import Ema (
 import Ema.CLI qualified
 import Ema.Dynamic (Dynamic (Dynamic))
 import Emanote.CLI qualified as CLI
+import Emanote.Model.Graph qualified as G
 import Emanote.Model.Link.Rel (ResolvedRelTarget (..))
 import Emanote.Model.Type (modelCompileTailwind)
 import Emanote.Model.Type qualified as Model
@@ -51,8 +52,17 @@ instance IsRoute SiteRoute where
 
 instance EmaSite SiteRoute where
   type SiteArg SiteRoute = EmanoteConfig
-  siteInput = emanoteSiteInput
+  siteInput cliAct cfg = do
+    model <- emanoteSiteInput cliAct cfg
+    pure $ model <&> modelUpdateCachedFields
   siteOutput = View.emanoteSiteOutput
+
+-- | Populate model fields that needs to be computed once per update.
+modelUpdateCachedFields :: Model.ModelEma -> Model.ModelEma
+modelUpdateCachedFields model =
+  model
+    & Model.modelFolgezettelTree
+    .~ G.folgezettelTreesFrom (unModelEma model) (Model.modelIndexRoute model)
 
 defaultEmanoteConfig :: CLI.Cli -> EmanoteConfig
 defaultEmanoteConfig cli =
@@ -67,8 +77,8 @@ run cfg@EmanoteConfig {..} = do
         >>= postRun cfg
     CLI.Cmd_Export -> do
       Dynamic (unModelEma -> model0, _) <-
-        flip runLoggerLoggingT oneOffLogger $
-          siteInput @SiteRoute (Ema.CLI.action def) cfg
+        flip runLoggerLoggingT oneOffLogger
+          $ siteInput @SiteRoute (Ema.CLI.action def) cfg
       putLBSLn $ Export.renderJSONExport model0
   where
     -- A logger suited for running one-off commands.
@@ -84,8 +94,8 @@ run cfg@EmanoteConfig {..} = do
 
 postRun :: EmanoteConfig -> (Model.ModelEma, (FilePath, [FilePath])) -> IO ()
 postRun EmanoteConfig {..} (unModelEma -> model0, (outPath, genPaths)) = do
-  when (model0 ^. modelCompileTailwind) $
-    compileTailwindCss (outPath </> generatedCssFile) genPaths
+  when (model0 ^. modelCompileTailwind)
+    $ compileTailwindCss (outPath </> generatedCssFile) genPaths
   checkBrokenLinks _emanoteConfigCli $ Export.modelRels model0
   checkBadMarkdownFiles $ Model.modelNoteErrors model0
 
@@ -104,8 +114,9 @@ checkBadMarkdownFiles noteErrs = runStderrLoggingT $ do
 
 checkBrokenLinks :: CLI.Cli -> Map LMLRoute [Export.Link] -> IO ()
 checkBrokenLinks cli modelRels = runStderrLoggingT $ do
-  ((), res :: Sum Int) <- runWriterT $
-    forM_ (Map.toList modelRels) $ \(noteRoute, rels) ->
+  ((), res :: Sum Int) <- runWriterT
+    $ forM_ (Map.toList modelRels)
+    $ \(noteRoute, rels) ->
       forM_ (sortNub rels) $ \(Export.Link urt rrt) ->
         case rrt of
           RRTFound _ -> pass
@@ -127,11 +138,15 @@ compileTailwindCss :: (MonadUnliftIO m) => FilePath -> [FilePath] -> m ()
 compileTailwindCss cssPath genPaths = do
   runStdoutLoggingT $ do
     log $ "Running Tailwind CSS v3 compiler to generate: " <> toText cssPath
-    Tailwind.runTailwind $
-      def
-        & Tailwind.tailwindConfig % Tailwind.tailwindConfigContent .~ genPaths
-        & Tailwind.tailwindOutput .~ cssPath
-        & Tailwind.tailwindMode .~ Tailwind.Production
+    Tailwind.runTailwind
+      $ def
+      & Tailwind.tailwindConfig
+      % Tailwind.tailwindConfigContent
+      .~ genPaths
+      & Tailwind.tailwindOutput
+      .~ cssPath
+      & Tailwind.tailwindMode
+      .~ Tailwind.Production
 
 defaultEmanotePandocRenderers :: EmanotePandocRenderers Model.Model LMLRoute
 defaultEmanotePandocRenderers =