+++ emanote/src/Emanote/Route/R.hs
@@ -22,8 +22,10 @@ instance (HasExt ext) => ToJSON (R ext) where
 
 instance (HasExt ext) => Show (R ext) where
   show r =
-    toString $
-      "R[/" <> encodeRoute r <> "]"
+    toString
+      $ "R[/"
+      <> encodeRoute r
+      <> "]"
 
 -- | Convert foo/bar.<ext> to a @R@
 mkRouteFromFilePath :: forall a (ext :: FileType a). (HasExt ext) => FilePath -> Maybe (R ext)
@@ -52,7 +54,7 @@ routeSlugWithPrefix prefix r = do
 -- | The base name of the route without its parent path.
 routeBaseName :: R ext -> Text
 routeBaseName =
-  Slug.unSlug . head . NE.reverse . unRoute
+  Slug.unSlug . last . unRoute
 
 routeParent :: R ext -> Maybe (R 'Folder)
 routeParent =