+++ emanote/src/Emanote/View/TaskIndex.hs
@@ -23,10 +23,12 @@ newtype TaskIndex = TaskIndex {unTaskIndex :: Map R.LMLRoute (NonEmpty Task)}
 
 mkTaskIndex :: Model -> TaskIndex
 mkTaskIndex model =
-  TaskIndex . Map.map NE.sort $
-    Map.fromListWith (<>) $
-      filter (not . Task._taskChecked) (Ix.toList $ model ^. M.modelTasks) <&> \task ->
-        (task ^. Task.taskRoute, one task)
+  TaskIndex
+    . Map.map NE.sort
+    $ Map.fromListWith (<>)
+    $ filter (not . Task._taskChecked) (Ix.toList $ model ^. M.modelTasks)
+    <&> \task ->
+      (task ^. Task.taskRoute, one task)
 
 renderTasks :: Model -> LByteString
 renderTasks model = do
@@ -42,8 +44,9 @@ renderTasks model = do
       taskSplice task = do
         let r = task ^. Task.taskRoute
         -- TODO: reuse note splice
-        "task:description" ## Common.withInlineCtx tCtx $ \ctx ->
-          Splices.pandocSplice ctx $ B.Pandoc mempty $ one $ B.Plain $ task ^. Task.taskDescription
+        "task:description" ##
+          Common.withInlineCtx tCtx $ \ctx ->
+            Splices.pandocSplice ctx $ B.Pandoc mempty $ one $ B.Plain $ task ^. Task.taskDescription
         "note:title" ## Common.titleSplice tCtx (M.modelLookupTitle r model)
         "note:url" ## HI.textSplice (SR.siteRouteUrl model $ SR.lmlSiteRoute (R.LMLView_Html, r))